Added recursive route loading and refactored code
