+++ package.json
@@ -12,6 +12,8 @@
     "body-parser": "^1.19.0",
     "dotenv": "^9.0.0",
     "express": "^4.17.1",
+    "fs": "*",
+    "path": "^0.12.7",
     "sqlite3": "^5.0.2"
   },
   "devDependencies": {
