+++ routes/users/getAllUsers.js
@@ -1,0 +1,16 @@
+const db = require.main.require('./database');
+
+module.exports = function(app) {
+    app.get('/users', (req, res) => {
+        db.all('select * from users', [], function (err, rows) {
+            if (err) {
+                res.status(400).json({ error: err.message });
+                return;
+            }
+            res.json({
+                ok: true,
+                users: rows
+            });
+        });
+    });
+}
